added wallet system

+++ be/java-sb/src/main/java/com/pfm/sbjwt/controllers/TestController.java
@@ -5,8 +5,9 @@ import com.pfm.sbjwt.models.Asset;
 import com.pfm.sbjwt.models.ExchangeRate;
 import com.pfm.sbjwt.models.Expense;
 import com.pfm.sbjwt.models.Income;
-import com.pfm.sbjwt.models.InitialState;
 import com.pfm.sbjwt.models.User;
+import com.pfm.sbjwt.models.WalletEntry;
+import com.pfm.sbjwt.models.WalletOrigin;
 import com.pfm.sbjwt.payload.request.AddAssetRequest;
 import com.pfm.sbjwt.payload.request.AddExpenseRequest;
 import com.pfm.sbjwt.payload.request.AddIncomeRequest;
@@ -23,21 +24,25 @@ import com.pfm.sbjwt.payload.response.UserResponse;
 import com.pfm.sbjwt.payload.response.models.AssetNetwork;
 import com.pfm.sbjwt.payload.response.models.ExpenseNetwork;
 import com.pfm.sbjwt.payload.response.models.IncomeNetwork;
-import com.pfm.sbjwt.payload.response.models.InitialStateNetwork;
+import com.pfm.sbjwt.payload.response.models.WalletNetwork;
 import com.pfm.sbjwt.repository.AssetRepository;
 import com.pfm.sbjwt.repository.ExpenseRepository;
 import com.pfm.sbjwt.repository.IncomeRepository;
-import com.pfm.sbjwt.repository.InitialStateRepository;
 import com.pfm.sbjwt.repository.UserRepository;
+import com.pfm.sbjwt.repository.WalletDetailsRepository;
+import com.pfm.sbjwt.repository.WalletOriginRepository;
 import com.pfm.sbjwt.security.jwt.JwtUtils;
 import com.pfm.sbjwt.services.ExchangeRateService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.Arrays;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import org.antlr.v4.runtime.misc.Pair;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -61,7 +66,9 @@ public class TestController {
 
   private final UserRepository userRepository;
 
-  private final InitialStateRepository initialStateRepository;
+  private final WalletDetailsRepository walletDetailsRepository;
+
+  private final WalletOriginRepository walletOriginRepository;
 
   private final ExpenseRepository expenseRepository;
 
@@ -75,7 +82,8 @@ public class TestController {
       ExchangeRateUpdater exchangeRateUpdater,
       JwtUtils jwtUtils,
       UserRepository userRepository,
-      InitialStateRepository initialStateRepository,
+      WalletDetailsRepository walletDetailsRepository,
+      WalletOriginRepository walletOriginRepository,
       ExpenseRepository expenseRepository,
       IncomeRepository incomeRepository,
       AssetRepository assetRepository) {
@@ -83,7 +91,8 @@ public class TestController {
     this.exchangeRateUpdater = exchangeRateUpdater;
     this.jwtUtils = jwtUtils;
     this.userRepository = userRepository;
-    this.initialStateRepository = initialStateRepository;
+    this.walletDetailsRepository = walletDetailsRepository;
+    this.walletOriginRepository = walletOriginRepository;
     this.expenseRepository = expenseRepository;
     this.incomeRepository = incomeRepository;
     this.assetRepository = assetRepository;
@@ -103,6 +112,17 @@ public class TestController {
       exchangeRates = exchangeRateUpdater.updateExchangeRates();
     }
 
+    // Check if the wallet must be prolonged to a new month
+    Optional<LocalDate> latestWalletDate = walletDetailsRepository.getLatestDateByUser(user);
+    LocalDate today = LocalDate.now();
+    if (latestWalletDate.isPresent() && latestWalletDate.get().getMonth() != today.getMonth()) {
+      LocalDate newDate = today.minusDays(today.getDayOfMonth() - 1);
+      List<WalletEntry> prolongedEntries =
+          walletDetailsRepository.getWalletEntriesByDateEqualsAndUser(latestWalletDate.get(), user);
+      prolongedEntries.forEach(
+          entry -> walletDetailsRepository.save(new WalletEntry(newDate, entry)));
+    }
+
     return ResponseEntity.ok(new UserResponse(exchangeRates, user));
   }
 
@@ -110,61 +130,128 @@ public class TestController {
   @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
   public ResponseEntity<?> userSetup(
       @Valid @RequestBody SetupRequest setupRequest, HttpServletRequest request) {
-    User user = getUserFromRequest(request);
+    final User user = getUserFromRequest(request);
     if (user == null) {
       return ResponseEntity.badRequest().body(new MessageResponse("Can't find user data!"));
     }
 
-    BigDecimal amount = setupRequest.getValue();
-    LocalDate startDate = setupRequest.getStartDate();
-    InitialState initialState = user.getInitialState();
-    if (initialState != null) {
-      initialStateRepository.modifyInitialStateById(initialState.getId(), startDate, amount);
-      initialState.setValues(startDate, amount);
+    final LocalDate newStartDate = setupRequest.getStartDate();
+    final List<WalletEntry> newWalletEntries =
+        setupRequest.getEntries().entrySet().stream()
+            .map(entry -> new WalletEntry(user, newStartDate, entry.getKey(), entry.getValue()))
+            .toList();
+    final List<WalletEntry> prevWalletEntries = user.getWalletEntries();
+
+    List<WalletEntry> finalWalletEntries;
+    if (!prevWalletEntries.isEmpty()) {
+      // A wallet was already set, need to modify it, for every timespan affected
+      LocalDate prevStartDate =
+          prevWalletEntries.stream()
+              .map(WalletEntry::getDate)
+              .min(LocalDate::compareTo)
+              .orElseThrow();
+
+      finalWalletEntries = new ArrayList<>();
+      for (LocalDate date = newStartDate; !date.isEqual(prevStartDate); date = date.plusMonths(1)) {
+        final LocalDate finalDate = date;
+        finalWalletEntries.addAll(
+            newWalletEntries.stream().map(entry -> new WalletEntry(finalDate, entry)).toList());
+      }
+      // Save the new ones
+      if (!finalWalletEntries.isEmpty()) {
+        walletDetailsRepository.saveAll(finalWalletEntries);
+      }
+
+      // Modify the previous ones
+      // Compute the offset w.r.t old origin points
+      List<WalletOrigin> prevOrigins = user.getWalletOrigin();
+      Map<String, BigDecimal> newOrigins =
+          WalletNetwork.groupByDate(newWalletEntries).get(newStartDate);
+      Map<String, Float> offsets = new HashMap<>();
+      for (WalletOrigin origin : prevOrigins) {
+        offsets.put(
+            origin.getCurrencyCode(),
+            newOrigins.get(origin.getCurrencyCode()).floatValue()
+                - origin.getAmount().floatValue());
+      }
+
+      Map<LocalDate, Map<String, Pair<BigDecimal, Long>>> groupedEntries =
+          WalletNetwork.groupByDateWithId(prevWalletEntries);
+      for (Map.Entry<LocalDate, Map<String, Pair<BigDecimal, Long>>> entry :
+          groupedEntries.entrySet()) {
+        entry
+            .getValue()
+            .forEach(
+                (key, value) -> {
+                  BigDecimal newValue = BigDecimal.valueOf(value.a.floatValue() + offsets.get(key));
+                  walletDetailsRepository.modifyEntryById(value.b, newValue);
+                  finalWalletEntries.add(new WalletEntry(user, entry.getKey(), key, newValue));
+                });
+      }
+      walletOriginRepository.deleteAllByUser(user);
     } else {
-      // Set the amount for the username
-      initialState = new InitialState(user, amount, startDate);
-      initialStateRepository.save(initialState);
+      // Ensure to fill up until today
+      List<WalletEntry> prolongedEntries = new ArrayList<>();
+      LocalDate firstOfMonth = LocalDate.now().minusDays(LocalDate.now().getDayOfMonth() - 1);
+      for (LocalDate date = newStartDate.plusMonths(1);
+          !date.isAfter(firstOfMonth);
+          date = date.plusMonths(1)) {
+        LocalDate finalDate = date;
+        newWalletEntries.forEach(entry -> prolongedEntries.add(new WalletEntry(finalDate, entry)));
+      }
+      finalWalletEntries = new ArrayList<>(newWalletEntries);
+      finalWalletEntries.addAll(prolongedEntries);
+      walletDetailsRepository.saveAll(finalWalletEntries);
     }
 
-    return ResponseEntity.ok(new SetupResponse(new InitialStateNetwork(initialState)));
+    walletOriginRepository.saveAll(newWalletEntries.stream().map(WalletOrigin::new).toList());
+    return ResponseEntity.ok(
+        new SetupResponse(new WalletNetwork(finalWalletEntries, newStartDate)));
   }
 
   @PostMapping("/user/expense/add")
   @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
   public ResponseEntity<?> userExpenseAdd(
-      @Valid @RequestBody AddExpenseRequest[] addExpenseRequests, HttpServletRequest request) {
+      @Valid @RequestBody AddExpenseRequest addExpenseRequest, HttpServletRequest request) {
     User user = getUserFromRequest(request);
     if (user == null) {
       return ResponseEntity.badRequest().body(new MessageResponse("Can't find user data!"));
     }
 
     // Get the expense from the request
-    List<Expense> expenses =
-        Arrays.stream(addExpenseRequests)
-            .map(addExpenseRequest -> addExpenseRequest.buildExpense(user))
-            .toList();
-    expenseRepository.saveAll(expenses);
+    Expense expense = addExpenseRequest.buildExpense(user);
+    expenseRepository.save(expense);
 
-    return ResponseEntity.ok(new ExpenseResponse(expenses));
+    Entry entry =
+        new Entry(
+            addExpenseRequest.getCurrencyCode(),
+            -addExpenseRequest.getAmount(),
+            addExpenseRequest.getDate());
+    WalletNetwork newWallet = modifyWallet(user, entry);
+
+    return ResponseEntity.ok(new ExpenseResponse(expense, newWallet));
   }
 
   @PostMapping("/user/income/add")
   @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
   public ResponseEntity<?> userIncomeAdd(
-      @Valid @RequestBody AddIncomeRequest[] addIncomeRequests, HttpServletRequest request) {
+      @Valid @RequestBody AddIncomeRequest addIncomeRequest, HttpServletRequest request) {
     User user = getUserFromRequest(request);
     if (user == null) {
       return ResponseEntity.badRequest().body(new MessageResponse("Can't find user data!"));
     }
 
-    List<Income> incomes =
-        Arrays.stream(addIncomeRequests)
-            .map(addIncomeRequest -> addIncomeRequest.buildIncome(user))
-            .toList();
-    incomeRepository.saveAll(incomes);
+    Income income = addIncomeRequest.buildIncome(user);
+    incomeRepository.save(income);
 
-    return ResponseEntity.ok(new IncomeResponse(incomes));
+    Entry entry =
+        new Entry(
+            addIncomeRequest.getCurrencyCode(),
+            addIncomeRequest.getAmount(),
+            addIncomeRequest.getDate());
+    WalletNetwork newWallet = modifyWallet(user, entry);
+
+    return ResponseEntity.ok(new IncomeResponse(income, newWallet));
   }
 
   @PostMapping("/user/asset/add")
@@ -179,20 +266,33 @@ public class TestController {
     Asset asset = addAssetRequest.buildAsset(user);
     assetRepository.save(asset);
 
-    return ResponseEntity.ok(new AssetResponse(new AssetNetwork(asset)));
+    Float offset = addAssetRequest.getToBePurchased() ? 0.f : asset.getAmount().floatValue();
+    // Modify the wallet with the new asset
+    WalletNetwork newWallet =
+        modifyWallet(user, new Entry(asset.getCurrencyCode(), offset, addAssetRequest.getDate()));
+
+    return ResponseEntity.ok(new AssetResponse(new AssetNetwork(asset), newWallet));
   }
 
   @PostMapping("/user/expense/modify")
   @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
   public ResponseEntity<?> userExpenseModify(
-      @Valid @RequestBody ModifyExpenseRequest modifyExpenseRequest) {
-    ExpenseNetwork expense;
+      @Valid @RequestBody ModifyExpenseRequest modifyExpenseRequest, HttpServletRequest request) {
+    User user = getUserFromRequest(request);
+    if (user == null) {
+      return ResponseEntity.badRequest().body(new MessageResponse("Can't find user data!"));
+    }
+
+    ExpenseNetwork expense = modifyExpenseRequest.getExpense();
+    Entry entry;
     if (modifyExpenseRequest.getDelete()) {
-      Long id = modifyExpenseRequest.getExpense().getId();
+      Long id = expense.getId();
       expenseRepository.removeExpenseById(id);
+      entry =
+          new Entry(expense.getCurrencyCode(), expense.getAmount().floatValue(), expense.getDate());
       expense = new ExpenseNetwork(id);
     } else {
-      expense = modifyExpenseRequest.getExpense();
+      BigDecimal oldAmount = expenseRepository.getExpenseById(expense.getId()).getAmount();
       expenseRepository.modifyExpenseById(
           expense.getId(),
           expense.getDate(),
@@ -201,22 +301,37 @@ public class TestController {
           expense.getSubCategory(),
           expense.getDescription(),
           expense.getAmount());
+      entry =
+          new Entry(
+              expense.getCurrencyCode(),
+              -expense.getAmount().floatValue() + oldAmount.floatValue(),
+              expense.getDate());
     }
 
-    return ResponseEntity.ok(new ExpenseResponse(expense));
+    WalletNetwork newWallet = modifyWallet(user, entry);
+
+    return ResponseEntity.ok(new ExpenseResponse(expense, newWallet));
   }
 
   @PostMapping("/user/income/modify")
   @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
   public ResponseEntity<?> userIncomeModify(
-      @Valid @RequestBody ModifyIncomeRequest modifyIncomeRequest) {
-    IncomeNetwork income;
+      @Valid @RequestBody ModifyIncomeRequest modifyIncomeRequest, HttpServletRequest request) {
+    User user = getUserFromRequest(request);
+    if (user == null) {
+      return ResponseEntity.badRequest().body(new MessageResponse("Can't find user data!"));
+    }
+
+    IncomeNetwork income = modifyIncomeRequest.getIncome();
+    Entry entry;
     if (modifyIncomeRequest.getDelete()) {
-      Long id = modifyIncomeRequest.getIncome().getId();
+      Long id = income.getId();
       incomeRepository.removeIncomeById(id);
+      entry =
+          new Entry(income.getCurrencyCode(), -income.getAmount().floatValue(), income.getDate());
       income = new IncomeNetwork(id);
     } else {
-      income = modifyIncomeRequest.getIncome();
+      BigDecimal oldAmount = incomeRepository.getIncomeById(income.getId()).getAmount();
       incomeRepository.modifyIncomeById(
           income.getId(),
           income.getDate(),
@@ -225,21 +340,38 @@ public class TestController {
           income.getSubCategory(),
           income.getDescription(),
           income.getAmount());
+      entry =
+          new Entry(
+              income.getCurrencyCode(),
+              income.getAmount().floatValue() - oldAmount.floatValue(),
+              income.getDate());
     }
 
-    return ResponseEntity.ok(new IncomeResponse(income));
+    WalletNetwork newWallet = modifyWallet(user, entry);
+
+    return ResponseEntity.ok(new IncomeResponse(income, newWallet));
   }
 
   @PostMapping("/user/asset/modify")
   @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') or hasRole('ADMIN')")
   public ResponseEntity<?> userAssetModify(
-      @Valid @RequestBody ModifyAssetRequest modifyAssetRequest) {
-    AssetNetwork asset;
+      @Valid @RequestBody ModifyAssetRequest modifyAssetRequest, HttpServletRequest request) {
+    User user = getUserFromRequest(request);
+    if (user == null) {
+      return ResponseEntity.badRequest().body(new MessageResponse("Can't find user data!"));
+    }
+
+    AssetNetwork asset = modifyAssetRequest.getAsset();
+    Asset oldAsset = assetRepository.getAssetById(asset.getId());
+    Entry entry;
     if (modifyAssetRequest.getDelete()) {
-      Long id = modifyAssetRequest.getAsset().getId();
+      Long id = asset.getId();
       assetRepository.removeAssetById(id);
+      entry = new Entry(asset.getCurrencyCode(), -asset.getAmount().floatValue(), asset.getDate());
       asset = new AssetNetwork(id);
     } else {
+      // Consider the floating prices later
+      BigDecimal oldAmount = oldAsset.getAmount();
       asset = modifyAssetRequest.getAsset();
       assetRepository.modifyAssetById(
           asset.getId(),
@@ -249,9 +381,42 @@ public class TestController {
           asset.getDescription(),
           asset.getIdentifierCode(),
           asset.getAmount());
+      entry =
+          new Entry(
+              asset.getCurrencyCode(),
+              asset.getAmount().floatValue() - oldAmount.floatValue(),
+              asset.getDate());
     }
 
-    return ResponseEntity.ok(new AssetResponse(asset));
+    WalletNetwork newWallet = modifyWallet(user, entry);
+
+    return ResponseEntity.ok(new AssetResponse(asset, newWallet));
+  }
+
+  private WalletNetwork modifyWallet(User user, Entry transaction) {
+    List<WalletEntry> newWalletEntries = new ArrayList<>();
+
+    Map<LocalDate, Map<String, Pair<BigDecimal, Long>>> oldWalletEntries =
+        WalletNetwork.groupByDateWithId(user.getWalletEntries());
+
+    Float currentOffset = transaction.getOffset();
+
+    // Dates are sorted in the map (tree map)
+    for (Map.Entry<String, Pair<BigDecimal, Long>> entry :
+        oldWalletEntries.get(transaction.getDate()).entrySet()) {
+      WalletEntry walletEntry;
+      if (entry.getKey().equals(transaction.getCurrency())) {
+        BigDecimal newAmount = BigDecimal.valueOf(entry.getValue().a.floatValue() + currentOffset);
+        walletDetailsRepository.modifyEntryById(entry.getValue().b, newAmount);
+        walletEntry = new WalletEntry(user, transaction.getDate(), entry.getKey(), newAmount);
+      } else {
+        walletEntry =
+            new WalletEntry(user, transaction.getDate(), entry.getKey(), entry.getValue().a);
+      }
+      newWalletEntries.add(walletEntry);
+    }
+
+    return new WalletNetwork(newWalletEntries);
   }
 
   private User getUserFromRequest(HttpServletRequest request) {
@@ -262,5 +427,30 @@ public class TestController {
       return null;
     }
     return user.get();
+  }
+
+  private static class Entry {
+    private final String currency;
+    private final Float offset;
+
+    private final LocalDate date;
+
+    public Entry(String currency, Float offset, LocalDate date) {
+      this.currency = currency;
+      this.offset = offset;
+      this.date = date;
+    }
+
+    public String getCurrency() {
+      return currency;
+    }
+
+    public Float getOffset() {
+      return offset;
+    }
+
+    public LocalDate getDate() {
+      return date;
+    }
   }
 }
